updates to calendar/events
