+++ src/App.js
@@ -1,5 +1,5 @@
 import routes from './routes'
-import Header from './Components/Header'
+import Header from './Components/Header/Header'
 import './App.css';
 
 function App() {
