+++ src/Components/Calendar/MyCalendar.css
@@ -8,7 +8,17 @@
     --bg-color: #1D3557;
     --neutral-color: #1D3557;
   }
-
+  
+  /* .sizing {
+    background-color: #1D3557;
+    width: 50%;
+    height: 50%;
+    align-items: right;
+  } */
+  /* .container {
+    background-color: #1D3557;
+  } */
+  
   html,
   body,
   body > div { /* the react root */
@@ -20,7 +30,7 @@
   
   h2 {
     margin: 0;
-    font-size: 16px;
+    font-size: 14px;
     text-transform: capitalize;
   }
   
@@ -41,13 +51,14 @@
   
   .demo-app {
     display: flex;
-    min-height: 100%;
+    min-height: 50%;
     font-family: Arial, Helvetica Neue, Helvetica, sans-serif;
-    font-size: 14px;
+    font-size: 12px;
     text-transform: capitalize;
   }
   
   .demo-app-sidebar {
+    margin: 100px 0 0 0;
     width: 300px;
     line-height: 1.5;
     background: #1D3557;
@@ -65,10 +76,11 @@
     flex-grow: 1;
     padding: 3em;
     background-color: #F1FAEE;
+    font-size: 14px;
     color: #457B9D;
     font-family: Arial, Helvetica, sans-serif;
     text-transform: capitalize;
-    margin: 0 2rem 3rem 0;
+    /* margin: .5em 0 10em 0; */
   }
   
   .fc {
