--- src/Components/Events/EventAgenda.js
@@ -1,6 +1,0 @@
-function EventAgenda({ event }) {
-    return <span>
-    <em>{event.title}</em>
-    <p>{event.desc}</p>
-  </span>
-}
