--- src/Components/Events/Events.css
@@ -1,3 +1,0 @@
-h1 {
-    padding: 30px;
-}
